Sandbox-SBX-2035/Likes and favorites add React.memo, add README.md
