+++ README.md
@@ -1,0 +1,33 @@
+# 📰 inFocus
+
+inFocus — это одностраничное приложение для просмотра новостей с возможностью лайков, авторизации под разными пользователями и ведением статистики.  
+
+## Функционал
+- Псевдоавторизация — выбор одного из нескольких пользователей.
+- Лайки новостей — возможность лайкать/анлайкать карточки.
+- История лайков — отмена и возврат действия.
+- Избранное — фильтрация новостей по лайкам.
+- Адаптивный интерфейс с анимациями.
+
+## Технологии
+- **React** — основа фронтенда  
+- **Express (Node.js)** — бэкенд для авторизации  
+- **SASS** — стилизация интерфейса  
+- **Webpack** — сборка проекта  
+
+## Установка и запуск
+- **Node.js**: >= 23.x (рекомендуемая версия для стабильной работы)  
+- **npm**: >= 10.x  
+```bash
+# Клонирование репозитория
+git clone https://github.com/username/newspeek.git
+
+# Установка зависимостей
+npm install
+
+# Запуск клиента
+webpack serve --mode development --open
+
+# Запуск сервера (отдельно)
+cd worker-mock-server
+npx ts-node server.ts
