--- news-mock-server/server.ts
@@ -1,44 +1,0 @@
-import express, { Request, Response } from 'express';
-import cors from 'cors';
-import fs from 'fs';
-import path from 'path';
-
-const app = express();
-const PORT = 3000;
-
-app.use('/images', express.static(path.join(__dirname, 'public/images')));
-app.use(cors());
-app.use(express.json());
-
-const dataPath = path.join(__dirname, 'data', 'news.json');
-
-interface NewsItem {
-  id: number;
-  image: string;
-  title: string;
-  tags: string[];
-  description: string;
-  content: string;
-  date: string;
-  author: string;
-}
-
-interface NewsData { news: NewsItem[]; }
-
-app.get('/news', (req: Request, res: Response) => {
-  const page = parseInt(req.query.page as string) || 1;
-  const limit = parseInt(req.query.limit as string) || 4;
-  const start = (page - 1) * limit;
-  const end = start + limit;
-  const raw = fs.readFileSync(dataPath, 'utf-8');
-  const jsonNews: NewsData = JSON.parse(raw);
-  const sliced = jsonNews.news.slice(start, end);
-  res.json({
-    news: sliced,
-    hasMore: end < jsonNews.news.length
-  });
-});
-
-app.listen(PORT, () => {
-  console.log(`Запущен: http://localhost:${PORT}`);
-});
